Trabalhando na requisição do pedido
Basicamente estou criando uma lógica que permitirá com que a tabela de itensPedido do banco seja preenchido automaticamente sempre que um novo pedido for feito.

+++ src/main/java/Controller/Montador/MontadorPedido.java
@@ -18,9 +18,9 @@ public class MontadorPedido implements MontadorReadAll<Pedido>{
     @Override
     public Pedido montar(ResultSet rs) throws SQLException {
        Pedido ped = new Pedido();
-       ped.setId_pedido(rs.getLong("id_pedido"));
-       ped.setId_requisitante(rs.getLong("id_requisitante"));
-       ped.setId_funcionario(rs.getLong("id_funcionario"));
+       ped.setId_pedido(rs.getInt("id_pedido"));
+       ped.setId_requisitante(rs.getInt("id_requisitante"));
+       ped.setId_funcionario(rs.getInt("id_funcionario"));
        return ped;
     }
