+++ src/main/java/Utils/Utilitarios.java
@@ -11,13 +11,13 @@ import java.util.concurrent.ThreadLocalRandom;
  * @author Isaias
  */
 public class Utilitarios {
-         public static long gerar_id(String  tipo){//verificar dentro do banco se o id ja existe
-                 long id = switch (tipo) {
-                     case "produto" -> ThreadLocalRandom.current().nextLong(100000000, 999999999);//length=9
+         public static Integer gerar_id(String  tipo){//verificar dentro do banco se o id ja existe
+                 Integer id = switch (tipo) {
+                     case "produto" -> ThreadLocalRandom.current().nextInt(100000000, 999999999);//length=9
                      case "pedido" ->
-                             ThreadLocalRandom.current().nextLong(10000000, 99999999);//length =8
-                     case "funcionario" ->    ThreadLocalRandom.current().nextLong(1000000, 9999999);//length =7
-                     case "empresa" -> ThreadLocalRandom.current().nextLong(100000, 999999);//length = 6
+                             ThreadLocalRandom.current().nextInt(10000000, 99999999);//length =8
+                     case "funcionario" ->    ThreadLocalRandom.current().nextInt(1000000, 9999999);//length =7
+                     case "empresa" -> ThreadLocalRandom.current().nextInt(100000, 999999);//length = 6
                      default -> throw new IllegalArgumentException("Tipo de ID desconhecido: " + tipo);
                  };
                  //verificar se o id gerado já existe. se for o caso, gerar outro
